requirements.txt
streamlit
langchain
langchain-google-genai
langchain-community
PyPDF2
faiss-cpu
python-dotenv
google-generativeai
